#1.1 Validation for unique name on add to DB, Update and Find items by ID in the DB, Delete by ID.
